fix(onboarding): close final review regressions
