+++ tests/conventions/i18n-key-resolution.test.ts
@@ -35,9 +35,12 @@ export const DYNAMIC_KEY_SITES = [
   "app/[locale]/(protected)/inbox/components/thread-state-picker.tsx :: t :: Inbox.threadStates.",
   "app/[locale]/(protected)/onboarding/wizard/components/onboarding-wizard.tsx :: t :: OnboardingWizard.steps.",
   "app/[locale]/(protected)/profile/components/api-key-modal.tsx :: t :: OnboardingWizard.ai.choices.",
+  "components/ai-connection/ai-connection-api-key-setup.tsx :: t :: OnboardingWizard.ai.choices.",
+  "components/ai-connection/ai-connection-api-key-setup.tsx :: t :: OnboardingWizard.ai.install.instruction.",
+  "components/ai-connection/ai-connection-claude-setup.tsx :: t :: OnboardingWizard.ai.choices.",
+  "components/ai-connection/ai-connection-claude-setup.tsx :: t :: OnboardingWizard.ai.methods.",
   "components/ai-connection/ai-connection-flow.tsx :: t :: OnboardingWizard.ai.choices.",
-  "components/ai-connection/ai-connection-flow.tsx :: t :: OnboardingWizard.ai.install.instruction.",
-  "components/ai-connection/ai-connection-flow.tsx :: t :: OnboardingWizard.ai.methods.",
+  "components/ai-connection/ai-connection-provider-grid.tsx :: t :: OnboardingWizard.ai.choices.",
   "app/[locale]/(protected)/profile/components/account-status-color.ts :: t :: Common.providers.",
   "app/[locale]/(protected)/profile/components/connected-account-modal.tsx :: t :: ConnectedAccountsCard.statusLabels.",
   "app/[locale]/(protected)/profile/components/connected-accounts-card.tsx :: t :: ConnectedAccountsCard.statusLabels.",
